Introduce intermediary SearchResult class
