+++ CHANGELOG.md
@@ -28,7 +28,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 - Improve time taken by Github actions by using cache ([#439](https://github.com/opensearch-project/opensearch-java/pull/439))
-- SearchTemplateResponse now extends SearchResponse, making all expected response fields accessible ([#462](https://github.com/opensearch-project/opensearch-java/pull/462))
+- Introduce intermediary SearchResult for SearchResponse and SearchTemplateResponse classes, enabling similar response handling for both ([#462](https://github.com/opensearch-project/opensearch-java/pull/462))
 
 ### Deprecated
